+++ src/main.zig
@@ -1,6 +1,6 @@
 const std = @import("std");
+const init_checks = @import("services/init_checks.zig");
 
 pub fn main() !void {
-    // Prints to stderr (it's a shortcut based on `std.io.getStdErr()`)
-    std.debug.print("All your {s} are belong to us.\n", .{"codebase"});
+    try init_checks.init_checks();
 }