+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ### Added 9/9/2024
 
 - Initial release
+- Added init_checks.zig that checks for the database etc before starting the app
+- Added config.zig with all the config variables
+- Created logger util for better logging
+- Created path alias on build.zig for easier imports